+++ src/com/qagwaai/starmalaccamax/server/dao/objectify/ObjectifyStarDAO.java
@@ -34,7 +34,7 @@ public class ObjectifyStarDAO implements StarDAO {
 			if (criterion instanceof SimpleFilterItem) {
 				SimpleFilterItem item = (SimpleFilterItem) criterion;
 				if (item.getField().equals("solarSystemId")) {
-					command = command.filter(item.getField(), Integer.valueOf(item.getValue()).intValue());
+					command = command.filter(item.getField(), Long.valueOf(item.getValue()).longValue());
 				} else if ( item.getField().equals("id")) {
 					command = (Query<StarDTO>) command.filterKey(Key.create(StarDTO.class, Long.valueOf(item.getValue()).longValue()));
 				}
